Add files via upload
Project was to create a room booking system in whatever language we desired. We decided to code the project in python using the python GUI tkinter. Various project management topics were learned and we learnt how to accept change requests openly. Even though this was a very basic project using waterfall method and not common AGILE practices, it was an eye opener to coding larger projects in python and learning the Software development cycle